+++ lib/flutter_pom.dart
@@ -3,6 +3,7 @@ library flutter_pom;
 export 'model/table.dart';
 export 'model/field.dart';
 export 'model/database.dart';
+export 'model/serializable.dart';
 
 // Fields
 
@@ -13,6 +14,7 @@ export 'model/fields/double_field.dart';
 export 'model/fields/datetime_field.dart';
 export 'model/fields/bool_field.dart';
 export 'model/fields/secure_string_field.dart';
+export 'model/fields/object_field.dart';
 
 // Errors