+++ lib/model/serializable.dart
@@ -1,0 +1,4 @@
+abstract class Serializable {
+  Map<String, String> toJson();
+  void fromJson(String value);
+}